// src/types/theme/theme.type.ts
export interface IColorPalette {
  main: string;
  lightPurple: string;
  darkPurple: string;
  btnGray: string;
  lightGray: string;
  white: string;
  lightBlack: string;
  black: string;
}

export interface IFontPalette {
  font14: string;
  font16: string;
  font18: string;
  font20: string;
  font22: string;
  font24: string;
  font28: string;
  font30: string;
  font36: string;
  font40: string;
  font48: string;
  font60: string;
}

export interface ItemType {
  picture: string;
  type: boolean;
  like: string;
  chat: number;
  poster: string;
  albumTitle: string;
}